requirements.txt
pandas
numpy
keras
seaborn
Pillow
tensorflow
tensorflow-gpu
matplotlib
albumentations
scikit-learn
